Commit lesson_status changes
If handleCompletionStatus is set to true and completion status is set to either "not attempted" or "unknown", lesson_status is automatically set to "incomplete". However, scorm.save() (aka Commit) was not being invoked, potentially causing the changes to be lost. scorm.save() is now invoked, ensuring changes are persisted in LMS. Hat tip to Jørgen Nielsen